DOS source
In order to compile Heks in DOS, I used Turbo C, by Borland International.
of course, since that compiler was old and incompatible with modern versions of C, I did need to adept the code a bit in order to enable Turbo C to compile everything accordingly.
Now I do not expect any serious bugs in any version of Heks, but please note that the DOS version has less priority to be up-to-date than the other versions. I suppose that's obvious...